Update readme.md
It occurred to me this morning that we don’t need to support the JSON object payload in the node-pptx lib itself. This should be done at the microservice level. Then we’ll just iterate over the object and create the pptx as needed. Not adding this logic will keep the lib itself smaller.